# carkee-api/src/Swagger/swagger.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server",
      "variables": {}
    },
    {
      "url": "https://carkee-staging.solveware.co",
      "description": "Live server",
      "variables": {}
    }
  ],
  "info": {
    "version": "b238a7c0-2985-4770-9d8d-f1ea0152611f",
    "title": "Carkee",
    "description": "",
    "termsOfService": "",
    "contact": {},
    "license": {
      "name": ""
    }
  },
  "paths": {
    "/api/user/register": {
      "post": {
        "summary": "Registration",
        "operationId": "Registration",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Seccessfully execution of API",
            "headers": {}
          },
          "402": {
            "description": "Data not found"
          },
          "400": {
            "description": "Kindly share rquired fields"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "password_confirm",
                  "account_id",
                  "fullname",
                  "mobile_code",
                  "mobile"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "harry99@yopmail.or"
                  },
                  "password": {
                    "type": "string",
                    "example": "Pass!234"
                  },
                  "password_confirm": {
                    "type": "string",
                    "example": "Pass!234"
                  },
                  "account_id": {
                    "type": "string",
                    "example": "1"
                  },
                  "fullname": {
                    "type": "string",
                    "example": "test remo"
                  },
                  "mobile_code": {
                    "type": "string",
                    "example": "+65"
                  },
                  "mobile": {
                    "type": "string",
                    "example": "882874300"
                  }
                }
              },
              "example": [
                {
                  "key": "email",
                  "value": "harry99@yopmail.or",
                  "type": "text"
                },
                {
                  "key": "password",
                  "value": "Pass!234",
                  "type": "text"
                },
                {
                  "key": "password_confirm",
                  "value": "Pass!234",
                  "type": "text"
                },
                {
                  "key": "account_id",
                  "value": "1",
                  "type": "text"
                },
                {
                  "key": "fullname",
                  "value": "test remo",
                  "type": "text"
                },
                {
                  "key": "mobile_code",
                  "value": "+65",
                  "type": "text"
                },
                {
                  "key": "mobile",
                  "value": "882874300",
                  "type": "text"
                }
              ]
            }
          }
        },
        "tags": ["User Registration"]
      }
    },
...
    "/api/ads/createAds": {
      "post": {
        "summary": "create Ads",
        "description": "api/ads -- create ads",
        "operationId": "createAds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Seccessfully execution of API",
            "headers": {}
          },
          "402": {
            "description": "Data not found"
          },
          "400": {
            "description": "Kindly share rquired fields"
          }
        },
        "requestBody": {
          "description": "api/ads -- create ads",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "description", "link", "isBottom", "file"],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "carkee ads new"
                  },
                  "description": {
                    "type": "string",
                    "example": "ads for carkee app 2"
                  },
                  "link": {
                    "type": "string",
                    "example": "https://adslink2.com"
                  },
                  "isBottom": {
                    "type": "string",
                    "example": "1"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "example": [
                {
                  "key": "name",
                  "value": "carkee ads new",
                  "type": "text"
                },
                {
                  "key": "description",
                  "value": "ads for carkee app 2",
                  "type": "text"
                },
                {
                  "key": "link",
                  "value": "https://adslink2.com",
                  "type": "text"
                },
                {
                  "key": "isBottom",
                  "value": "1",
                  "type": "text"
                },
                {
                  "key": "file",
                  "type": "file",
                  "src": []
                }
              ]
            }
          }
        },
        "tags": ["Ads"]
      }
    },
...
    "/api/ads/updateAds": {
      "patch": {
        "summary": "update ads",
        "description": "api/ads/update -- update ads using ads Id,give ads Id in the request body",
        "operationId": "updateads",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Seccessfully execution of API",
            "headers": {}
          },
          "402": {
            "description": "Data not found"
          },
          "400": {
            "description": "Kindly share rquired fields"
          }
        },
        "requestBody": {
          "description": "api/ads/update -- update ads using ads Id,give ads Id in the request body",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description",
                  "link",
                  "isBottom",
                  "file",
                  "Ads_id"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "carkee ads updated"
                  },
                  "description": {
                    "type": "string",
                    "example": "ads for carkee app 2"
                  },
                  "link": {
                    "type": "string",
                    "example": "https://adslink2.com"
                  },
                  "isBottom": {
                    "type": "string",
                    "example": "1"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "Ads_id": {
                    "type": "string",
                    "example": "16"
                  }
                }
              },
              "example": [
                {
                  "key": "name",
                  "value": "carkee ads updated",
                  "type": "text"
                },
                {
                  "key": "description",
                  "value": "ads for carkee app 2",
                  "type": "text"
                },
                {
                  "key": "link",
                  "value": "https://adslink2.com",
                  "type": "text"
                },
                {
                  "key": "isBottom",
                  "value": "1",
                  "type": "text"
                },
                {
                  "key": "file",
                  "type": "file",
                  "src": "/home/prem/Downloads/blackjean.jpeg"
                },
                {
                  "key": "Ads_id",
                  "value": "16",
                  "type": "text"
                }
              ]
            }
          }
        },
        "tags": ["Ads"]
      }
    },
    "/api/ads/deleteAds": {
      "delete": {
        "summary": "Remove Ads",
        "description": "remove ads ",
        "operationId": "RemoveAds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Seccessfully execution of API",
            "headers": {}
          },
          "402": {
            "description": "Data not found"
          },
          "400": {
            "description": "Kindly share rquired fields"
          }
        },
        "tags": ["Ads"]
      }
    },
    "/api/ads/updateisBottom": {
      "patch": {
        "summary": "updateisBottom",
        "operationId": "updateisBottom",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Seccessfully execution of API",
            "headers": {}
          },
          "402": {
            "description": "Data not found"
          },
          "400": {
            "description": "Kindly share rquired fields"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "text/plain": {
              "schema": {
                "type": "string",
                "example": {
                  "Ads_id": 8,
                  "isBottom": 0
                }
              },
              "example": "{\n    \"Ads_id\" : 8,\n    \"isBottom\" : 0\n}"
            }
          }
        },
        "tags": ["Ads"]
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "noauth": {
        "type": "http",
        "scheme": "noauth"
      }
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
      "bearer": []
    },
    {
...
}